load model weights as f16 except rmsnorm, which is as f32

+++ src/main.rs
@@ -1,4 +1,5 @@
 use anyhow::Result;
+use dfdx::dtypes::f16;
 use dfdx::prelude::*;
 use rand::prelude::{SeedableRng, StdRng};
 
@@ -88,7 +89,7 @@ fn main() -> Result<()> {
 
     let cfg = ConfigV2 {};
 
-    let m = TinyLlama::<f32, _, _>::load_model(cfg, &dev, &loader)?;
+    let m = TinyLlama::<f16, _, _>::load_model(cfg, &dev, &loader)?;
 
     let gen_opt = GenerateOption {
         use_cache: !args.disable_cache,
